backend/requirements.txt
transformers==4.30.2
torch==2.0.1
Pillow==9.5.0
TTS==0.13.3
numpy==1.24.3 